requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ftfy
regex
tqdm
Pillow
numpy
clip @ git+https://github.com/openai/CLIP.git
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0 